WIP Update Android NDK to r16
All the hardcoding

Sister CLs:
src/third_party/skia/: https://skia-review.googlesource.com/c/skia/+/75422
src/third_party/android_tools/ndk/: https://chromium-review.googlesource.com/c/android_ndk/+/784230

Bug: 771171
Cq-Include-Trybots: master.tryserver.chromium.android:android_optional_gpu_tests_rel;master.tryserver.chromium.linux:linux_optional_gpu_tests_rel;master.tryserver.chromium.mac:mac_optional_gpu_tests_rel;master.tryserver.chromium.win:win_optional_gpu_tests_rel

+++ base/android/linker/legacy_linker_jni.cc
@@ -93,7 +93,7 @@ bool GenericLoadLibrary(JNIEnv* env,
   crazy_context_t* context = GetCrazyContext();
 
   if (!IsValidAddress(load_address)) {
-    LOG_ERROR("Invalid address 0x%llx", load_address);
+    LOG_ERROR("Invalid address 0x%llx", (unsigned long long)load_address);
     return false;
   }
 
@@ -328,7 +328,7 @@ jboolean CreateSharedRelro(JNIEnv* env,
   LOG_INFO("Called for %s", lib_name.c_str());
 
   if (!IsValidAddress(load_address)) {
-    LOG_ERROR("Invalid address 0x%llx", load_address);
+    LOG_ERROR("Invalid address 0x%llx", (unsigned long long)load_address);
     return false;
   }
